Moved briefcase.toml path handling to BaseCommand.
